refactor: 각 클래스 책임 분리
