+++ apps/web/ios/App/App/PushNotificationHandler.swift
@@ -1,0 +1,49 @@
+import UIKit
+import FirebaseMessaging
+import UserNotifications
+
+final class PushNotificationHandler: NSObject, UNUserNotificationCenterDelegate {
+    func configure(application: UIApplication) {
+        UNUserNotificationCenter.current().delegate = self
+        application.registerForRemoteNotifications()
+    }
+
+    func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        willPresent notification: UNNotification,
+        withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void
+    ) {
+        let userInfo = notification.request.content.userInfo
+        Messaging.messaging().appDidReceiveMessage(userInfo)
+        completionHandler([.alert, .sound, .badge])
+    }
+
+    func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        didReceive response: UNNotificationResponse,
+        withCompletionHandler completionHandler: @escaping () -> Void
+    ) {
+        let userInfo = response.notification.request.content.userInfo
+        Messaging.messaging().appDidReceiveMessage(userInfo)
+        completionHandler()
+    }
+
+    func didRegisterForRemoteNotifications(with deviceToken: Data) {
+        print("📍 APNs 등록 성공! 토큰 생성 시작...")
+        Messaging.messaging().apnsToken = deviceToken
+        Messaging.messaging().token(completion: { (token, error) in
+            if let error = error {
+                NotificationCenter.default.post(name: .capacitorDidFailToRegisterForRemoteNotifications, object: error)
+                print("FCM 토큰 에러: \(error)\(error.localizedDescription)")
+            } else if let token = token {
+                NotificationCenter.default.post(name: .capacitorDidRegisterForRemoteNotifications, object: token)
+                print("FCM 토큰: \(token)")
+            }
+        })
+    }
+
+    func didFailToRegisterForRemoteNotifications(with error: Error) {
+        print("❌ APNs 등록 실패: \(error.localizedDescription)")
+        NotificationCenter.default.post(name: .capacitorDidFailToRegisterForRemoteNotifications, object: error)
+    }
+}
